requirements.txt
flask
gunicorn
httpx
scrapy
flask-restful
python-dotenv